REFACTOR relative paths after move
